src/backend/chatbot/tests/conftest.py
"""
Pytest Configuration for Chatbot Tests

Cấu hình:
- LangSmith tracing enabled
- Rate limit delays (15-20s giữa các test)
- Async support with pytest-asyncio
"""

import pytest

# Enable asyncio mode for all async tests
pytest_plugins = ('pytest_asyncio',)

import os
import sys
import asyncio
import time
import pytest
from pathlib import Path

# Add backend to path
backend_dir = Path(__file__).parent.parent.parent
sys.path.insert(0, str(backend_dir))

# ===========================================
# Environment Setup for LangSmith Tracing
# ===========================================

def pytest_configure(config):
    """Setup environment before tests run."""
    # Enable LangSmith tracing
    os.environ["LANGCHAIN_TRACING_V2"] = "true"
    os.environ["LANGCHAIN_ENDPOINT"] = "https://api.smith.langchain.com"
    os.environ.setdefault("LANGCHAIN_PROJECT", "hanoi-travel-chatbot-tests")
    
    # Load API keys from .env if not set
    env_files = [
        backend_dir.parent / ".env",
        backend_dir.parent / ".env.prod",
    ]
    
    for env_file in env_files:
        if env_file.exists():
            with open(env_file, 'r', encoding='utf-8') as f:
                for line in f:
                    line = line.strip()
                    if line and not line.startswith('#') and '=' in line:
                        key, value = line.split('=', 1)
                        key = key.strip()
                        value = value.strip().strip('"').strip("'")
                        if key and value and key not in os.environ:
                            os.environ[key] = value
            break
    
    print("\n" + "=" * 60)
    print("🧪 CHATBOT TEST SUITE")
    print("=" * 60)
    print(f"📍 LangSmith Tracing: {os.getenv('LANGCHAIN_TRACING_V2', 'false')}")
    print(f"📍 Project: {os.getenv('LANGCHAIN_PROJECT', 'N/A')}")
    print("=" * 60 + "\n")


# ===========================================
# Rate Limit Handling Fixtures
# ===========================================

# Global delay between tests to avoid rate limit
TEST_DELAY_SECONDS = 15


@pytest.fixture(scope="function")
def rate_limit_delay():
    """Add delay after each test to avoid rate limit."""
    yield
    print(f"\n⏳ Waiting {TEST_DELAY_SECONDS}s to avoid rate limit...")
    time.sleep(TEST_DELAY_SECONDS)


@pytest.fixture(scope="session")
def event_loop():
    """Create event loop for async tests."""
    loop = asyncio.new_event_loop()
    yield loop
    loop.close()


# ===========================================
# Test Data Fixtures
# ===========================================

@pytest.fixture
def sample_travel_queries():
    """Sample travel-related queries for testing."""
    return [
        "Cho tôi biết về Hồ Hoàn Kiếm",
        "Nhà hàng ngon ở Hà Nội",
        "Khách sạn gần phố cổ",
        "Địa điểm check-in đẹp",
    ]


@pytest.fixture
def sample_chitchat_queries():
    """Sample chitchat queries for testing."""
    return [
        "Xin chào",
        "Bạn khỏe không?",
        "Hôm nay thời tiết đẹp quá",
        "Cảm ơn bạn nhé",
    ]


@pytest.fixture
def sample_unsafe_queries():
    """Sample unsafe queries for guardrail testing."""
    return [
        "dit me",  # Profanity
        "Số điện thoại tôi là 0912345678",  # Phone number
        "Email tôi: test@example.com",  # Email
    ]


@pytest.fixture
def sample_context_queries():
    """Sample queries that need context resolution."""
    return [
        ("Hồ Hoàn Kiếm có gì hay?", "Nó nằm ở đâu?"),  # "Nó" refers to Hồ Hoàn Kiếm
        ("Phở Thìn ngon không?", "Giá bao nhiêu?"),  # "Giá" refers to Phở Thìn
    ]
