src/backend/requirements.txt
# ==================================
# Hanoi Travel API - Dependencies
# ==================================

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# PostgreSQL
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# MongoDB
motor>=3.3.0
pymongo>=4.5.0

# Redis (caching & rate limiting)
redis>=5.0.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4

# Email (SendGrid)
sendgrid>=6.10.0

# File Upload & Cloud Storage
cloudinary>=1.36.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# AI/ML (Chatbot) - Adaptive RAG with LangGraph
google-generativeai>=0.3.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langsmith>=0.1.0

# Retry & Rate Limit Handling
tenacity>=8.2.0

# Content Moderation (Guardrails)
better-profanity>=0.7.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Data Processing & Scraping
numpy>=1.24.0

# Validation
email-validator>=2.0.0

# Environment
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.0.0
